refactor: move the db cleaner to the bootstrap file
